rename blockie -> blocky
