+++ design-system/icons/Blocky.svelte
